Updated /profile, added ui settings to user preferences
