+++ src/management-system-v2/components/content.module.scss
@@ -24,7 +24,6 @@
     overflow-y: auto;
     padding: 10px 10px;
     max-width: 100%;
-    background-color: white;
     height: calc(100vh - 150px);
 
     &.compact {
