--- src/management-system-v2/components/userProfile.module.scss
@@ -1,43 +1,0 @@
-.Row {
-  cursor: pointer;
-
-  .Title {
-    font-weight: 600;
-  }
-
-  .ActionCell {
-    display: flex;
-    justify-content: center;
-
-    .Action {
-      background-color: #ececec;
-      border-radius: 100%;
-      padding: 4px;
-      font-size: 17px;
-      transition: background-color 0.2s;
-    }
-
-    .Action:hover {
-      background-color: #dcdcdc;
-    }
-  }
-}
-
-.Card {
-  width: min(600px, 90%);
-  margin-left: 3vw;
-
-  .Buttons {
-    display: flex;
-    @media screen and (max-width: 550px) {
-      flex-direction: column;
-      align-items: start;
-    }
-  }
-}
-
-@media screen and (max-width: 800px) {
-  .Content {
-    margin-left: 0px;
-  }
-}
